pyproject.toml
[project]
name = "localstack-learn"
version = "0.1.0"
description = "AWS learning project with LocalStack emulation"
requires-python = ">=3.11"

dependencies = [
    # AWS SDK
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    
    # LocalStack
    "localstack>=3.0.0",
    
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    
    # Environment management
    "python-dotenv>=1.0.0",
    
    # HTTP client for testing
    "requests>=2.31.0",
    
    # CLI tools
    "click>=8.1.0",
    "rich>=13.7.0",  # Beautiful CLI output
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.uv]
# UV configuration

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
